--- tests/behat/features/bootstrap/MongoDB_GridFS.php
@@ -1,22 +1,0 @@
-<?php
-/**
- * This file is part of the Imbo package
- *
- * (c) Christer Edvartsen <cogo@starzinger.net>
- *
- * For the full copyright and license information, please view the LICENSE file that was
- * distributed with this source code.
- */
-
-namespace ImboBehatFeatureContext;
-
-/**
- * Imbo Context
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
- * @package Test suite\Functional tests
- */
-class MongoDB_GridFS extends MainFeatureContext implements FeatureContext {
-    use Database\MongoDB;
-    use Storage\GridFS;
-}
